Update mainNav and other
