+++ lang/en.json
@@ -54,5 +54,7 @@
     "Kokybiškos studijos ir joms pritaikyta aplinka": "High quality studies and their environment",
     "Stipri organizacija" : "Strong organization",
     "Darni universitetinė bendruomenė" : "Sustainable University community",
-    "Padėka": "Acknowledgements"
+    "Padėka": "Acknowledgements",
+    "Klausykis studentiško": "Listen to students'",
+    "radijo": "radio"
 }
