+++ lang/lt.json
@@ -57,5 +57,7 @@
     "Padėka": "Acknowledgements",
     "Too Many Requests": "Per daug užklausų",
     "Not Found": "Puslapis nerastas",
-    "Server Error": "Serverio klaida"
+    "Server Error": "Serverio klaida",
+    "Klausykis studentiško": "Klausykis studentiško",
+    "radijo": "radijo"
 }
