+++ resources/js/Components/AppLogo.vue
@@ -1,3 +1,3 @@
 <template>
-  <img class="w-24" src="/logos/vusa.lin.hor.svg" />
+  <img class="max-w-lg" src="/logos/vusa.lin.hor.svg" />
 </template>
