+++ resources/js/Components/Public/FullWidth/MainLinks.vue
@@ -1,0 +1,59 @@
+<template>
+  <div class="mx-8 mb-8 lg:mx-16 lg:px-16">
+    <h2 class="mb-4">Pagrindinės nuorodos:</h2>
+    <div class="flex flex-wrap gap-2">
+      <NButton
+        v-for="item in mainPage"
+        :key="item.id"
+        secondary
+        round
+        @click="goToLink(item.link)"
+      >
+        {{ item.text }}
+      </NButton>
+    </div>
+  </div>
+</template>
+
+<script setup lang="ts">
+import { Inertia } from "@inertiajs/inertia";
+import { NButton } from "naive-ui";
+import { usePage } from "@inertiajs/inertia-vue3";
+import route from "ziggy-js";
+
+defineProps<{ mainPage: Array<App.Models.MainPage> }>();
+
+const goToLink = (link: string | null) => {
+  // if link is null, return nothing
+  if (link === null) {
+    return;
+  }
+
+  // check if link is external
+  let padalinysAlias = usePage().props.value.alias;
+  if (link.includes("http")) {
+    window.open(link, "_blank");
+  }
+
+  // if has /lt/, truncate it
+  if (link.includes("/lt/")) {
+    link = link.replace("/lt/", "");
+  }
+
+  if (padalinysAlias === "vusa") {
+    // if first char is /, remove it
+    if (link.charAt(0) === "/") {
+      link = link.substring(1);
+    }
+    Inertia.visit(route("main.page", { lang: "lt", permalink: link }));
+  } else {
+    Inertia.visit(
+      route("padalinys.page", {
+        lang: "lt",
+        permalink: link,
+        padalinys: padalinysAlias,
+      })
+    );
+  }
+};
+</script>
