+++ resources/js/Components/Public/FullWidth/SiteFooter.vue
