+++ resources/js/Components/Public/Nav/FacebookButton.vue
@@ -1,0 +1,19 @@
+<template>
+  <NButton
+    text
+    @click="openBlank('https://www.facebook.com/VUstudentuatstovybe')"
+  >
+    <NIcon size="18">
+      <FacebookF />
+    </NIcon>
+  </NButton>
+</template>
+
+<script setup lang="ts">
+import { FacebookF } from "@vicons/fa";
+import { NButton, NIcon } from "naive-ui";
+
+const openBlank = (url: string) => {
+  window.open(url, "_blank");
+};
+</script>
