+++ resources/js/Pages/Public/Contacts/ContactsShow.vue
@@ -48,7 +48,7 @@ export default {
 import { Head } from "@inertiajs/inertia-vue3";
 
 import ContactWithPhotoForDuties from "@/Components/Public/ContactWithPhotoForDuties.vue";
-import FadeTransition from "@/Components/Public/FadeTransition.vue";
+import FadeTransition from "@/Components/Public/Utils/FadeTransition.vue";
 import ShapeDivider1 from "@/Components/Public/ShapeDivider1.vue";
 
 defineProps<{
