+++ resources/js/Pages/Public/NewsArchive.vue
@@ -85,7 +85,7 @@ import { NIcon, NPagination } from "naive-ui";
 import { ref } from "vue";
 import route from "ziggy-js";
 
-import FadeTransition from "@/Components/Public/FadeTransition.vue";
+import FadeTransition from "@/Components/Public/Utils/FadeTransition.vue";
 import HomeCard from "@/Components/Public/HomeCard.vue";
 
 const props = defineProps<{
