+++ resources/js/Pages/Public/SaziningaiExams.vue
@@ -116,7 +116,7 @@ import { h, ref } from "vue";
 import { useForm } from "@inertiajs/inertia-vue3";
 import route from "ziggy-js";
 
-import FadeTransition from "@/Components/Public/FadeTransition.vue";
+import FadeTransition from "@/Components/Public/Utils/FadeTransition.vue";
 
 const props = defineProps<{
   padaliniaiOptions: App.Models.Padalinys[];
